Switch score values from int to double

+++ app/src/androidTest/java/org/isoron/uhabits/widgets/views/CheckmarkWidgetViewTest.java
@@ -50,7 +50,7 @@ public class CheckmarkWidgetViewTest extends BaseViewTest
         habit = fixtures.createShortHabit();
         view = new CheckmarkWidgetView(targetContext);
         int color = ColorUtils.getAndroidTestColor(habit.getColor());
-        int score = habit.getScores().getTodayValue();
+        double score = habit.getScores().getTodayValue();
         float percentage = (float) score / Score.MAX_VALUE;
 
         view.setActiveColor(color);
